Move handlebars custom helpers to libimaginteraction
