+++ lib/core/libimagrt/src/logger.rs
@@ -35,6 +35,7 @@ use log::{Log, LogLevel, LogRecord, LogMetadata};
 use toml::Value;
 use toml_query::read::TomlValueReadExt;
 use handlebars::Handlebars;
+use libimaginteraction::format::*;
 
 type ModuleName = String;
 type Result<T> = ::std::result::Result<T, RE>;
@@ -82,19 +83,19 @@ impl ImagLogger {
 
         handlebars.register_escape_fn(::handlebars::no_escape);
 
-        handlebars.register_helper("black"  , Box::new(self::template_helpers::ColorizeBlackHelper));
-        handlebars.register_helper("blue"   , Box::new(self::template_helpers::ColorizeBlueHelper));
-        handlebars.register_helper("cyan"   , Box::new(self::template_helpers::ColorizeCyanHelper));
-        handlebars.register_helper("green"  , Box::new(self::template_helpers::ColorizeGreenHelper));
-        handlebars.register_helper("purple" , Box::new(self::template_helpers::ColorizePurpleHelper));
-        handlebars.register_helper("red"    , Box::new(self::template_helpers::ColorizeRedHelper));
-        handlebars.register_helper("white"  , Box::new(self::template_helpers::ColorizeWhiteHelper));
-        handlebars.register_helper("yellow" , Box::new(self::template_helpers::ColorizeYellowHelper));
+        handlebars.register_helper("black"  , Box::new(ColorizeBlackHelper));
+        handlebars.register_helper("blue"   , Box::new(ColorizeBlueHelper));
+        handlebars.register_helper("cyan"   , Box::new(ColorizeCyanHelper));
+        handlebars.register_helper("green"  , Box::new(ColorizeGreenHelper));
+        handlebars.register_helper("purple" , Box::new(ColorizePurpleHelper));
+        handlebars.register_helper("red"    , Box::new(ColorizeRedHelper));
+        handlebars.register_helper("white"  , Box::new(ColorizeWhiteHelper));
+        handlebars.register_helper("yellow" , Box::new(ColorizeYellowHelper));
 
-        handlebars.register_helper("underline"     , Box::new(self::template_helpers::UnderlineHelper));
-        handlebars.register_helper("bold"          , Box::new(self::template_helpers::BoldHelper));
-        handlebars.register_helper("blink"         , Box::new(self::template_helpers::BlinkHelper));
-        handlebars.register_helper("strikethrough" , Box::new(self::template_helpers::StrikethroughHelper));
+        handlebars.register_helper("underline"     , Box::new(UnderlineHelper));
+        handlebars.register_helper("bold"          , Box::new(BoldHelper));
+        handlebars.register_helper("blink"         , Box::new(BlinkHelper));
+        handlebars.register_helper("strikethrough" , Box::new(StrikethroughHelper));
 
         {
             let fmt = try!(aggregate_global_format_trace(matches, config));
@@ -476,143 +477,5 @@ fn aggregate_module_settings(_matches: &ArgMatches, config: Option<&Configuratio
             Ok(BTreeMap::new())
         }
     }
-}
-
-mod template_helpers {
-    use handlebars::{Handlebars, HelperDef, JsonRender, RenderError, RenderContext, Helper};
-    use ansi_term::Colour;
-    use ansi_term::Style;
-
-    #[derive(Clone, Copy)]
-    pub struct ColorizeBlackHelper;
-
-    impl HelperDef for ColorizeBlackHelper {
-        fn call(&self, h: &Helper, hb: &Handlebars, rc: &mut RenderContext) -> Result<(), RenderError> {
-            colorize(Colour::Black, h, hb, rc)
-        }
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct ColorizeBlueHelper;
-
-    impl HelperDef for ColorizeBlueHelper {
-        fn call(&self, h: &Helper, hb: &Handlebars, rc: &mut RenderContext) -> Result<(), RenderError> {
-            colorize(Colour::Blue, h, hb, rc)
-        }
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct ColorizeCyanHelper;
-
-    impl HelperDef for ColorizeCyanHelper {
-        fn call(&self, h: &Helper, hb: &Handlebars, rc: &mut RenderContext) -> Result<(), RenderError> {
-            colorize(Colour::Cyan, h, hb, rc)
-        }
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct ColorizeGreenHelper;
-
-    impl HelperDef for ColorizeGreenHelper {
-        fn call(&self, h: &Helper, hb: &Handlebars, rc: &mut RenderContext) -> Result<(), RenderError> {
-            colorize(Colour::Green, h, hb, rc)
-        }
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct ColorizePurpleHelper;
-
-    impl HelperDef for ColorizePurpleHelper {
-        fn call(&self, h: &Helper, hb: &Handlebars, rc: &mut RenderContext) -> Result<(), RenderError> {
-            colorize(Colour::Purple, h, hb, rc)
-        }
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct ColorizeRedHelper;
-
-    impl HelperDef for ColorizeRedHelper {
-        fn call(&self, h: &Helper, hb: &Handlebars, rc: &mut RenderContext) -> Result<(), RenderError> {
-            colorize(Colour::Red, h, hb, rc)
-        }
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct ColorizeWhiteHelper;
-
-    impl HelperDef for ColorizeWhiteHelper {
-        fn call(&self, h: &Helper, hb: &Handlebars, rc: &mut RenderContext) -> Result<(), RenderError> {
-            colorize(Colour::White, h, hb, rc)
-        }
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct ColorizeYellowHelper;
-
-    impl HelperDef for ColorizeYellowHelper {
-        fn call(&self, h: &Helper, hb: &Handlebars, rc: &mut RenderContext) -> Result<(), RenderError> {
-            colorize(Colour::Yellow, h, hb, rc)
-        }
-    }
-
-    fn colorize(color: Colour, h: &Helper, _: &Handlebars, rc: &mut RenderContext) -> Result<(), RenderError> {
-        let p = try!(h.param(0).ok_or(RenderError::new("Too few arguments")));
-
-        try!(write!(rc.writer(), "{}", color.paint(p.value().render())));
-        Ok(())
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct UnderlineHelper;
-
-    impl HelperDef for UnderlineHelper {
-        fn call(&self, h: &Helper, _: &Handlebars, rc: &mut RenderContext) -> Result<(),
-            RenderError> {
-                let p = try!(h.param(0).ok_or(RenderError::new("Too few arguments")));
-                let s = Style::new().underline();
-                try!(write!(rc.writer(), "{}", s.paint(p.value().render())));
-                Ok(())
-            }
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct BoldHelper;
-
-    impl HelperDef for BoldHelper {
-        fn call(&self, h: &Helper, _: &Handlebars, rc: &mut RenderContext) -> Result<(),
-            RenderError> {
-                let p = try!(h.param(0).ok_or(RenderError::new("Too few arguments")));
-                let s = Style::new().bold();
-                try!(write!(rc.writer(), "{}", s.paint(p.value().render())));
-                Ok(())
-            }
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct BlinkHelper;
-
-    impl HelperDef for BlinkHelper {
-        fn call(&self, h: &Helper, _: &Handlebars, rc: &mut RenderContext) -> Result<(),
-            RenderError> {
-                let p = try!(h.param(0).ok_or(RenderError::new("Too few arguments")));
-                let s = Style::new().blink();
-                try!(write!(rc.writer(), "{}", s.paint(p.value().render())));
-                Ok(())
-            }
-    }
-
-    #[derive(Clone, Copy)]
-    pub struct StrikethroughHelper;
-
-    impl HelperDef for StrikethroughHelper {
-        fn call(&self, h: &Helper, _: &Handlebars, rc: &mut RenderContext) -> Result<(),
-            RenderError> {
-                let p = try!(h.param(0).ok_or(RenderError::new("Too few arguments")));
-                let s = Style::new().strikethrough();
-                try!(write!(rc.writer(), "{}", s.paint(p.value().render())));
-                Ok(())
-            }
-    }
-
 }
