+++ lib/etc/libimaginteraction/Cargo.toml
@@ -22,6 +22,7 @@ log = "0.3"
 regex = "0.2"
 toml = "0.4"
 error-chain = "0.10"
+handlebars = "0.29.0"
 
 libimagstore       = { version = "0.5.0", path = "../../../lib/core/libimagstore" }
 libimagerror       = { version = "0.5.0", path = "../../../lib/core/libimagerror" }
